requirements.txt
numpy==1.20.1
Pillow==8.1.0
opencv-python==4.5.1.48
inputs==0.5
